requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.1
django-celery-results==2.4.0
django-ckeditor==6.5.0
django-cuser==2017.3.16
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-tinymce==3.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
modeltranslation==0.25
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.2.1
pywebio==1.6.3
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tornado==6.2
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
